Support for WebP and Gif output added (also quality option for PNG)
